+++ Changes.md
@@ -1,6 +1,7 @@
 HEAD
 -----------
 - Add support for running multiple Node Manager processes for added redundancy (#4)
+- Add support for specifying a redis database in RedisFailover::Client (#5)
 
 0.5.4
 -----------